+++ GAME/Game/Projectile.h
@@ -16,7 +16,7 @@ public:
 	void fire(btVector3 &pos, btVector3 &direction);
 	virtual void kill();
 	virtual void Initialize();	
-	virtual void Initialize(ISceneManager *smgr, BulletHelper *h, stringw projectileName, btVector3 &pos, btVector3 &dir, ElementalType eType, float dmg);
+	virtual void Initialize(ISceneManager *smgr, BulletHelper *h, stringw projectileName, btVector3 &pos, btVector3 &dir, ElementalType eType, int dmg);
 	virtual void Update(u32 deltaTime);	
 	virtual void revive();
 	ParticleManager::ParticleSystem *a;
@@ -24,7 +24,7 @@ public:
 	ParticleManager::ParticleSystem *c;
 	GameObjectType getType() const;
 	ElementalType getElementalType();
-	float getDamage();
+	int getDamage();
 	bool warmedUp;
 private:
 	ISceneManager *smgr;
